backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyMySQL==1.1.0
pandas>=2.2.3
scikit-learn==1.4.1.post1
numpy==1.26.4
django-cors-headers==4.3.1
python-dotenv==1.0.1
